Move some shared test objects into Shared.swift
