+++ Tests/appstoreconnect-cliTests/Operations/Shared.swift
@@ -1,0 +1,16 @@
+// Copyright 2020 Itty Bitty Apps Pty Ltd
+
+import Foundation
+
+let jsonDecoder: JSONDecoder = {
+    let jsonDecoder = JSONDecoder()
+    let dateFormatter = DateFormatter()
+    dateFormatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSSXXXXX"
+    jsonDecoder.dateDecodingStrategy = .formatted(dateFormatter)
+
+    return jsonDecoder
+}()
+
+enum TestError: Error, Equatable {
+    case somethingBadHappened
+}
